# prog/lab-4/python-glossary-api/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Python Glossary API",
    "description": "Глоссарий терминов Python: список, поиск, добавление, обновление и удаление терминов.",
    "version": "1.0.0"
  },
  "paths": {
    "/terms": {
      "get": {
        "summary": "Список всех терминов",
        "operationId": "list_terms_terms_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TermOut"
                  },
                  "type": "array",
                  "title": "Response List Terms Terms Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Добавить новый термин",
        "operationId": "create_term_terms_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TermCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TermOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/terms/{term}": {
      "get": {
        "summary": "Получить термин по ключевому слову",
        "operationId": "get_term_terms__term__get",
        "parameters": [
          {
            "name": "term",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Term"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TermOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Обновить существующий термин",
        "operationId": "update_term_terms__term__put",
        "parameters": [
          {
            "name": "term",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Term"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TermUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TermOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Удалить термин",
        "operationId": "delete_term_terms__term__delete",
        "parameters": [
          {
            "name": "term",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Term"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "TermCreate": {
        "properties": {
          "term": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Term",
            "examples": [
              "decorator"
            ]
          },
          "definition": {
            "type": "string",
            "minLength": 1,
            "title": "Definition",
            "examples": [
              "Функция, расширяющая поведение другой функции."
            ]
          },
          "category": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80
              },
              {
                "type": "null"
              }
            ],
            "title": "Category",
            "examples": [
              "Синтаксис"
            ]
          }
        },
        "type": "object",
        "required": [
          "term",
          "definition"
        ],
        "title": "TermCreate",
        "description": "Схема создания термина."
      },
      "TermOut": {
        "properties": {
          "term": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Term",
            "examples": [
              "decorator"
            ]
          },
          "definition": {
            "type": "string",
            "minLength": 1,
            "title": "Definition",
            "examples": [
              "Функция, расширяющая поведение другой функции."
            ]
          },
          "category": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80
              },
              {
                "type": "null"
              }
            ],
            "title": "Category",
            "examples": [
              "Синтаксис"
            ]
          },
          "id": {
            "type": "integer",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "term",
          "definition",
          "id"
        ],
        "title": "TermOut",
        "description": "Схема ответа."
      },
      "TermUpdate": {
        "properties": {
          "definition": {
            "type": "string",
            "minLength": 1,
            "title": "Definition"
          },
          "category": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          }
        },
        "type": "object",
        "required": [
          "definition"
        ],
        "title": "TermUpdate",
        "description": "Схема обновления (термин определяется по пути)."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}